requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
Pillow==10.1.0
requests==2.31.0
APScheduler==3.10.4
bcrypt==4.1.2
pytz==2023.3.post1
